Develop (#1)
* Modification du controlleur savedDates

* Modification de controllers

* Ajout de tests avec mockito

* Tests corrigés Mockito

* Ajout du format json aux routes

* Create LICENSE.md

* Modification du schéma des users dans la BDD

* Passage a jUnit4 et initialisation de mocks avant leur utilisation dans les tests

* Modif info compte

+++ src/main/java/CompteAr/backend/repository/SavedDatesRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SavedDatesRepository extends JpaRepository<SavedDates, Long> {
+public interface SavedDatesRepository extends JpaRepository<SavedDates, Integer> {
 
-        SavedDates findByUserId(long userId);
+        SavedDates findByUserId(Integer userId);
 
 
 }
